Split tmux session support declarations
